sandbox_202601catch/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "catch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13" # 3.12*にするために必要
dependencies = [
    "maturin",
    "numpy",
    "mujoco",
    "pillow>=12.1.0",
    "gymnasium>=1.2.3",
    "pytest",
    "torch",
    "torchvision",
    "stable-baselines3",
    "tensorboard",
    "tqdm",
    "rich",
    "moviepy",
]

[tool.uv]
package = true
# PyTorchの公式リポジトリを優先的に参照するように設定
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = { index = "pytorch-gpu" }